Created FileDetailForm component and added it in FileSelector component
